Add self-configuring CPU crash workaround script
Add "cpu-crash-workaround.sh" script to install the init service
"cpu_crash_workaround" that limit the CPU's maximum clock frequency to
1.0 GHz.  This works around stability issues with the IPQ8065 CPU on
the NBG6817.

Instructions:
1.  Copy this script to your router (SFTP/curl/wget)
2.  Make sure it's executable
# chmod +x cpu-crash-workaround.sh
3.  Install
# ./cpu-crash-workaround.sh install

To undo later:
# ./cpu-crash-workaround.sh remove

This automatically persists across sysupgrades by adding itself to the
backup list, INCLUDING backing up the fact that it's enabled.

NOTE: This will reduce performance!  It's only a workaround, NOT a
fix.  I created this to help ensure my two NBG6817 routers are stable
in between testing (one is at a remote location).